refactor: 뷰레코드 리포지토리 추상화
